requirements.txt
streamlit>=1.36.0
numpy>=1.26.0
plotly>=5.20.0
dimod>=0.12.0
dwave-neal>=0.6.0
